smart_ai_dashboard/requirements.txt
langchain
langchain_community
langgraph
groq
psycopg2
pandas
matplotlib
python-dotenv
